+++ src/app/home/challenge/[challengeId]/_componenets/DeleteChallengeModal.hook.tsx
@@ -1,0 +1,28 @@
+import { deleteChallenge } from "@/apis/services/challenge";
+import useModalStore from "@/stores/modalStore";
+import { useParams, useRouter } from "next/navigation";
+
+export default function useDeleteChallengeModal() {
+    const { closeModal } = useModalStore();
+    const { challengeId } = useParams();
+    const router = useRouter();
+
+    const handleCloseModal = () => {
+        closeModal();
+    };
+
+    const handleDelete = (e: React.FormEvent<HTMLFormElement>) => {
+        e.preventDefault();
+
+        deleteChallenge({ challengeId })
+            .then(() => {
+                closeModal();
+                router.push("/home/challenge");
+            })
+            .catch((error) => {
+                console.error(error);
+            });
+    };
+
+    return { handleDelete, handleCloseModal };
+}
